Not found error handling

+++ app/db/db_user.py
@@ -1,4 +1,6 @@
+from fastapi import HTTPException
 from sqlalchemy.orm import Session
+from starlette import status
 from app.db.hash import Hash
 from app.db.models import DbUser
 from app.schemas.schemas import UserBase
@@ -19,10 +21,15 @@ def get_all_users(db: Session):
     return db.query(DbUser).all()
 
 def get_user(db:Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User with id {id} not found')
+    return user
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User with id {id} not found')
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
@@ -32,7 +39,9 @@ def update_user(db: Session, id: int, request: UserBase):
     return 'ok'
 
 def delete_user(db: Session, id: int):
-    user = db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'User with id {id} not found')
     db.delete(user)
     db.commit()
     return 'ok'
